Aggiunta funzione free_node per tenere traccia dei nodi dell'heap liberati con la merge che altrimenti andrebbero persi. Modificata alloc_node in modo che ora usa se disponibili, i nodi scartati dalla merge Modificata la free

+++ include/libc/string.h
@@ -31,7 +31,7 @@
 int _kstrncmp (const char *, const char *, int);
 void *memset(void *, const int, int);
 extern void * memmove(void *,const void *,size_t);
-int strlen (const char *);
+size_t strlen (const char *);
 char *strncpy (char *, const char *, size_t);
 char *strchr (const char *, int);
 char *strstr (const char *, const char *);
